refactor and clean up post requests
