+++ core/src/services/onedrive/graph_model.rs
@@ -109,6 +109,40 @@ impl CreateDirPayload {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 struct EmptyStruct {}
 
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct FileUploadItem {
+    #[serde(rename = "@odata.type")]
+    odata_type: String,
+    #[serde(rename = "@microsoft.graph.conflictBehavior")]
+    microsoft_graph_conflict_behavior: String,
+    name: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OneDriveUploadSessionCreationResponseBody {
+    #[serde(rename = "uploadUrl")]
+    pub upload_url: String,
+    #[serde(rename = "expirationDateTime")]
+    pub expiration_date_time: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OneDriveUploadSessionCreationRequestBody {
+    item: FileUploadItem,
+}
+
+impl OneDriveUploadSessionCreationRequestBody {
+    pub fn new(path: String) -> Self {
+        OneDriveUploadSessionCreationRequestBody {
+            item: FileUploadItem {
+                odata_type: "microsoft.graph.driveItemUploadableProperties".to_string(),
+                microsoft_graph_conflict_behavior: "replace".to_string(),
+                name: path,
+            },
+        }
+    }
+}
+
 #[test]
 fn test_parse_one_drive_json() {
     let data = r#"{
